backend/requirements.txt
fastapi>=0.110.0,<0.116.0
uvicorn[standard]>=0.29.0
pydantic[email]>=2.5.0,<2.10
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.20.0
greenlet
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
openai>=1.0.0
sse-starlette>=1.6.0,<2.0.0
httpx>=0.27.0
apscheduler>=3.10.0
jinja2>=3.1.0
passlib[bcrypt]>=1.7.0
